+++ muna/beta/metadata/litert.py
@@ -15,6 +15,7 @@ class LiteRTInferenceMetadata(PyTorchInferenceMetadataBase):
     Members:
         model (torch.nn.Module): PyTorch module to apply metadata to.
         model_args (tuple[Tensor,...]): Positional inputs to the model.
+        input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
         output_keys (list): Model output dictionary keys. Use this if the model returns a dictionary.
     """
     kind: Literal["meta.inference.litert"] = Field(default="meta.inference.litert", init=False)